requirements.txt
streamlit==1.36.0

pymodm==0.4.3
pymongo==3.13.0
pymongo[srv]
bcrypt
python-dotenv